+++ src/com/yanncebron/m68kplugin/lang/psi/M68kMnemonic.java
@@ -102,7 +102,7 @@ public record M68kMnemonic(IElementType elementType,
     }
   }
 
-  enum ControlFlow {
+  public enum ControlFlow {
     NOTHING,
     TRAP,
     TRAP_RETURN,